Final Report
Extension. Ok so the CTE dataset looks at only injured rats and ignores sham rats, so we have 340 significant genes and clusters that have no obvious interpretation. Next iteration for each time point create a new dataset that contains genes that were significantly expressed between sham and injured, and then with this new dataset eliminate genes that were not significantly expressed overtime. I would bet that there would be less than 340 genes using this method, since it eliminates the genes that are expressed as a result of injury.